Fixed Thread Shut Down Bug
