+++ root/side_modules/logger.py
@@ -1,0 +1,7 @@
+import logging
+
+class Logger:
+    def __init__(self) -> None:    
+        logging.basicConfig()
+        self.logger = logging.getLogger('[SERVER]')
+        self.logger.setLevel(logging.INFO)
