+++ root/side_modules/settings.py
@@ -6,4 +6,4 @@ PORT = 5050
 HOST = 'localhost'
 ADDR = (HOST, PORT)
 FORMAT = 'utf-8'
-DISCONNECT_MESSAGE = '!DISCONNECT!'
+SHUT_DOWN_SERVER = False
